requirements.txt
requests>=2.31.0
PyYAML>=6.0
pytest>=7.4.0
pytest-mock>=3.11.0
responses>=0.23.0
